pyproject.toml
[project]
name = "nps-latam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "numpy>=2.2.1",
    "langchain-google-genai>=2.0.8",
    "langchain-core>=0.3.30",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.53.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "requests>=2.32.5",
    "evidently>=0.7.20",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pyzmq>=27.1.0",
    "jupyter>=1.0.0",
    "pytest>=9.0.2",
]
[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nps_latam"]
